+++ lib/SILOptimizer/Utils/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(UTILS_SOURCES
   Utils/CFG.cpp
+  Utils/CastOptimizer.cpp
   Utils/CheckedCastBrJumpThreading.cpp
   Utils/ConstantFolding.cpp
   Utils/Devirtualize.cpp
